clean up file structure
