+++ mac/keyrace-mac/KeyTap.swift
@@ -1,30 +1,8 @@
-//
-//  keyrace-mac
-//
-//  Created by Nat Friedman on 1/2/21.
-//
+// Created by Thomas Dohmke on 2/18/21.
 
-// TODO:
-// - Restructure, cleanup, learn swift
-// - Save username locally along with token
-// - Talk to server and show leaderboard
-
-
-import SwiftUI
 import Foundation
 import Accessibility
 import Cocoa
-
-@available(OSX 11.0, *)
-@main
-struct MenuBarPopoverApp: App {
-    @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
-    var body: some Scene {
-        Settings{
-            EmptyView()
-        }
-    }
-}
 
 func formatCount(count: Int) -> String {
     var str = ""
@@ -47,7 +25,7 @@ func formatCount(count: Int) -> String {
         default:
             pfx = ""
         }
-        
+
         var sfx = ""
         if (count < 100) {
             sfx = " today"
@@ -66,7 +44,7 @@ func myCGEventCallback(proxy: CGEventTapProxy, type: CGEventType, event: CGEvent
         var length = 0
         event.keyboardGetUnicodeString(maxStringLength: 1, actualStringLength: &length, unicodeString: &char)
         keyTap.increment(char)
-        
+
         keyTap.appDelegate.menubarItem?.statusBarItem.button?.title = formatCount(count: keyTap.keycount)
     }
 
@@ -88,15 +66,15 @@ class KeyTap {
     var minutes = [Int](repeating:0, count:1440)
     var keys = [Int](repeating: 0, count:256)
     var leaderboardText = NSMutableAttributedString()
-    
+
     init(_ appd: AppDelegate) {
         self.appDelegate = appd
     }
-    
+
     func increment(_ keyCode: UInt16) {
         let date = Date()
         let calendar = Calendar.current
-        
+
         // Reset to 0 at midnight
         let day = calendar.component(.day, from:date)
         if (lastDay != day) {
@@ -105,9 +83,9 @@ class KeyTap {
             minutes = [Int](repeating:0, count:1440)
             keys = [Int](repeating:0, count:256)
         }
-        
+
         keycount += 1
-        
+
         let hour = calendar.component(.hour, from: date)
         let minute = calendar.component(.minute, from: date)
         minutes[hour*60 + minute] += 1
@@ -119,10 +97,10 @@ class KeyTap {
             lastMin = minute
             uploadCount()
         }
-        
+
         saveCount()
     }
-    
+
     func getMinutesChart() -> [Int] {
         // Return the last 20 minutes minutely
         let date = Date()
@@ -139,7 +117,7 @@ class KeyTap {
         for i in 0..<minutes.count {
             hours[i/60] += minutes[i]
         }
-        
+
         return hours
     }
 
@@ -151,7 +129,7 @@ class KeyTap {
     func getLeaderboardText() -> NSMutableAttributedString {
         return leaderboardText
     }
-    
+
     func uploadKeycount() {
         if appDelegate.gh?.token == nil {
             return
@@ -168,7 +146,7 @@ class KeyTap {
 
         var request = URLRequest(url: url.url!)
         request.addValue("Bearer \(appDelegate.gh!.token!)", forHTTPHeaderField: "Authorization")
-        
+
         let task = URLSession.shared.dataTask(with: request) { data, response, error in
             guard let data = data,                            // is there data
                 let response = response as? HTTPURLResponse,  // is there HTTP response
@@ -178,20 +156,20 @@ class KeyTap {
                     return
             }
 
-            
+
             if let json_leaders = try? JSONSerialization.jsonObject(with: data, options: []) as? [[String: Any]] {
                 let title = "Leaderboard\n"
                 let attrTitle = NSMutableAttributedString(string: title)
                 attrTitle.addAttribute(.font, value: NSFont.monospacedSystemFont(ofSize: 14, weight: .bold) as Any, range: NSRange(location: 0, length: title.count))
                 self.leaderboardText = NSMutableAttributedString()
                 self.leaderboardText.append(attrTitle)
-                
+
                 // Add paragraph styling
                 let paragraphStyle = NSMutableParagraphStyle()
                 paragraphStyle.lineSpacing = 7
                 paragraphStyle.alignment = .justified
                 self.leaderboardText.addAttribute(.paragraphStyle, value: paragraphStyle, range: NSRange(location: 0, length: self.leaderboardText.length))
-                
+
                 for (i, leader) in json_leaders.enumerated(){
                     var u = ""
                     if let username = leader["username"] as? String {
@@ -212,7 +190,7 @@ class KeyTap {
                         }
                         s += "\n"
                     }
-                    
+
                     // Do the font styling for the line.
                     let attrLine = NSMutableAttributedString(string: fullUsername + s)
                     attrLine.addAttribute(.font, value: NSFont.monospacedSystemFont(ofSize: 12, weight: .regular), range: NSRange(location: 0, length: fullUsername.count + s.count))
@@ -224,21 +202,21 @@ class KeyTap {
                                               range: NSRange(location: 0, length: fullUsername.count))
                     self.leaderboardText.append(attrLine)
                 }
-                
+
                 self.leaderboardText.addAttribute(.paragraphStyle, value: paragraphStyle, range: NSRange(location: 0, length: self.leaderboardText.length))
                 self.leaderboardText.setAlignment(.justified, range: NSRange(location: 0, length: self.leaderboardText.length))
             }
-    
-           
+
+
         }
         task.resume()
     }
-    
+
     func setupKeyTap() {
         if (keyTrapSetup) {
             return
         }
-        
+
         loadCount()
 
         let eventMask = (1 << CGEventType.keyDown.rawValue)
@@ -259,14 +237,14 @@ class KeyTap {
         CGEvent.tapEnable(tap: eventTap, enable: true)
 
         appDelegate.menubarItem!.statusBarItem.button?.title = formatCount(count: keycount)
-        
+
         uploadCount()
     }
-    
+
     func uploadCount () {
         uploadKeycount()
     }
-    
+
     func saveCount() {
         var filename = FileManager.default.homeDirectoryForCurrentUser.appendingPathComponent(".keyrace.tmp")
 
@@ -277,7 +255,7 @@ class KeyTap {
             NSLog("Could not write keycount to \(filename.path)")
             // failed to write file – bad permissions, bad filename, missing permissions, or more likely it can't be converted to the encoding
         }
-        
+
         filename = FileManager.default.homeDirectoryForCurrentUser.appendingPathComponent(".keyrace.minutes.tmp")
         let minStrings = minutes.map({ String($0) })
         str = minStrings.joined(separator: ",")
@@ -299,7 +277,7 @@ class KeyTap {
         }
 
     }
-    
+
     func loadCount() {
         let date = Date()
         let calendar = Calendar.current
@@ -313,13 +291,13 @@ class KeyTap {
             let mtime = attr[FileAttributeKey.modificationDate] as! Date
             lastDay = calendar.component(.day, from:mtime)
             let lastMonth = calendar.component(.month, from:mtime)
-            
+
             if (lastDay == day && lastMonth == month) {
                 let str = try String(contentsOf: filename, encoding: .utf8)
                 keycount = Int(str) ?? 0
             }
         } catch { }
-        
+
         // Load the hourly histogram
         filename = FileManager.default.homeDirectoryForCurrentUser.appendingPathComponent(".keyrace.minutes.tmp")
         do {
@@ -360,7 +338,7 @@ class KeyTap {
             self.setupKeyTap()
             return
         }
-        
+
         // Wait for the user to give us permission
         if (timerRunning) { return }
         self.timerRunning = true
@@ -373,27 +351,5 @@ class KeyTap {
                 self.timerRunning = false
             }
         }
-    }
-}
-
-class AppDelegate: NSObject, NSApplicationDelegate {
-    var statusBarItem: NSStatusItem!
-    var menubarItem: MenubarItem?
-    var keyTap : KeyTap?
-    var gh : GitHub?
-    
-    func applicationDidFinishLaunching(_ notification: Notification) {
-        menubarItem = MenubarItem(title: "Setup Keyrace")
-
-        gh = GitHub()
-        menubarItem?.gh = gh
-
-        keyTap = KeyTap(self)
-        menubarItem?.keyTap = keyTap
-        keyTap!.getAccessibilityPermissions()
-    }
-    
-    func applicationWillTerminate(_ aNotification: Notification) {
-        // Insert code here to tear down your application
     }
 }
